+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 + Updated to forge 1.12.2-14.23.1.2596
 + Made spawner checking much more efficient
 + Added a config to filter spawning blocks
++ Added a particle to display spawn amount status